Fixed UI()->DoLabel with MaxWidth

+++ src/game/client/components/menus_manual/__defines_impl.h
@@ -17,14 +17,14 @@
 				char aBuf[512]; \
 				str_copy(aBuf, LINE, sizeof(aBuf)); \
 				aBuf[str_length(aBuf)-1] = '\0'; \
-				UI()->DoLabelScaled(&Label, aBuf+1, 20.0f, -1, (int)MainView.w); \
-				for(int i = 1; i < TextRender()->TextLineCount(0, 20.0f, LINE, (int)MainView.w); i++) MainView.HSplitTop(30.0f, &Label, &MainView); \
+				UI()->DoLabelScaled(&Label, aBuf+1, 20.0f, -1, MainView.w); \
+				for(int i = 1; i < TextRender()->TextLineCount(0, 20.0f, LINE, MainView.w); i++) MainView.HSplitTop(30.0f, &Label, &MainView); \
 			} \
 			else \
 			{ \
 				MainView.HSplitTop(20.0f, &Label, &MainView); \
-				UI()->DoLabelScaled(&Label, LINE, 15.f, -1, (int)(MainView.w - 30.0f)); \
-				for(int i = 1; i < TextRender()->TextLineCount(0, 15.0f, LINE, (int)(MainView.w - 30.0f)); i++) \
+				UI()->DoLabelScaled(&Label, LINE, 15.f, -1, (MainView.w - 30.0f)); \
+				for(int i = 1; i < TextRender()->TextLineCount(0, 15.0f, LINE, (MainView.w - 30.0f)); i++) \
 					MainView.HSplitTop(20.0f, &Label, &MainView); \
 			}
 #define NEWLINE() MainView.HSplitTop(20.0f, &Label, &MainView);
